requirements/base.txt
django==4.2
psycopg2-binary==2.9.6
python-environ==0.4.54
django-cors-headers==3.14.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-filter==23.1
django-rosetta==0.9.8
django-ckeditor==6.5.1
celery==5.2.7
pillow==9.5.0
redis==4.4.2
sentry-sdk==1.19.1
django-modeltranslation==0.18.9
python-telegram-bot==13.15  # for telegram bot integration
xlrd==2.0.1
